Auto save

+++ src/pages/notesPage/index.js
@@ -1,7 +1,18 @@
-import React, {useState} from 'react';
-import {View, Button, TextInput, StyleSheet} from 'react-native';
+import React, {useState, useEffect} from 'react';
+import {
+  View,
+  Button,
+  TextInput,
+  StyleSheet,
+  Alert,
+  ToastAndroid,
+  Platform,
+  Text,
+} from 'react-native';
+import {NavigationEvents} from 'react-navigation';
+import {TouchableOpacity} from 'react-native-gesture-handler';
 import useGlobal from '../../store';
-import {primaryColor, secondaryColor} from '../../config';
+import {secondaryColor} from '../../config';
 import debounce from '../../functions';
 
 const NotesPage = props => {
@@ -11,6 +22,12 @@ const NotesPage = props => {
   const [title, setTitle] = useState(navigation.getParam('title'));
   const [body, setBody] = useState(navigation.getParam('body'));
 
+  useEffect(() => {
+    navigation.setParams({
+      handleSave: callAddNoteAction,
+    });
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [title, body]);
   const inputValidation = () => {
     return title || body;
   };
@@ -29,16 +46,21 @@ const NotesPage = props => {
     } else {
       globalActions.editNote(newNote);
     }
+    if (Platform.OS === 'android') {
+      ToastAndroid.show('Note Saved!', ToastAndroid.SHORT);
+    }
     navigation.goBack();
   };
   return (
     <View style={styles.container}>
+      <NavigationEvents onWillBlur={() => callAddNoteAction()} />
       <TextInput
         style={styles.title}
         autoCapitalize="sentences"
         maxLength={25}
         onChangeText={value => debounce(setTitle(value), 300)}
         value={title}
+        placeholder="Title"
       />
       <TextInput
         multiline
@@ -48,21 +70,32 @@ const NotesPage = props => {
         style={styles.body}
         onChangeText={value => debounce(setBody(value), 300)}
         value={body}
+        placeholder="Type something Here"
       />
       <View style={styles.buttonHolder}>
-        <Button
-          onPress={() => {
-            callAddNoteAction(navigation);
-          }}
-          style={styles.button}
-          title="Save"
-          color={primaryColor}
-        />
         {id && (
           <Button
             onPress={() => {
-              globalActions.deleteNote(id);
-              navigation.goBack();
+              Alert.alert(
+                'Are you sure?',
+                '',
+                [
+                  {
+                    text: 'Delete',
+                    onPress: () => {
+                      globalActions.deleteNote(id);
+                      navigation.goBack();
+                    },
+                    style: 'destructive',
+                  },
+                  {
+                    text: 'Cancel',
+                    onPress: () => {},
+                    style: 'cancel',
+                  },
+                ],
+                {cancelable: false},
+              );
             }}
             style={styles.deleteButton}
             title="Delete"
@@ -72,6 +105,22 @@ const NotesPage = props => {
       </View>
     </View>
   );
+};
+
+NotesPage.navigationOptions = ({navigation}) => {
+  const {params = {}} = navigation.state;
+  console.log(navigation);
+  return {
+    headerRight: (
+      <TouchableOpacity
+        onPress={() => {
+          params.handleSave();
+        }}
+      >
+        <Text style={styles.saveButton}>Save</Text>
+      </TouchableOpacity>
+    ),
+  };
 };
 
 const styles = StyleSheet.create({
@@ -105,6 +154,11 @@ const styles = StyleSheet.create({
   },
   deleteButton: {
     margin: 10,
+  },
+  saveButton: {
+    fontSize: 20,
+    marginRight: 10,
+    color: '#fff',
   },
 });
